Record route and take picture properly
The app is able to map the route. It ignore the new coordinates if the difference between the previous coords and the new one is less 0.0001

Images can be taken from the camera properly.
Unfortunatly the promise didn't seem to work when calling navigator.camera.getPicture() function

NOTE: there is the occasional hiccups. The map takes a little bit of time to load. And taking picture button can take a little longer to execute. But on most cases it works fine

TODO: save the data properly.